拆分views_dependency
