+++ app/views_dependency.py
@@ -29,81 +29,12 @@ views
 
 @Date 2022-01-17
 '''
-from django.contrib.auth.decorators import login_required
 from django.http import HttpResponseRedirect
-from django.shortcuts import render
 from django.views.decorators.http import require_POST, require_GET
 
 from utils.hasher import MySHA256Hasher
 import utils.global_messages as my_messages
-from utils.global_messages import (
-    wrong,
-    succeed,
-    message_url,
-    append_query,
-)
-from utils.http.dependency import *
-from utils.views import *
-from generic.models import User
+from utils.global_messages import append_query
 from app.config import *
 from app import utils
-from app.log import logger
-
-
-# 不应导出，接口内部使用
-from django.core.exceptions import ImproperlyConfigured as _ImproperlyConfigured
-from app.log import ProfileLogger as _ProfileLogger
-
-class ProfileView(SecureView):
-    request: UserRequest
-    PrepareType = SecureView.NoReturnPrepareType | None
-
-    check_access: bool = True
-    need_prepare: bool = True
-    logger_name: str
-
-    def _check_access(self, redirect_url='/logout/') -> None:
-        if not self.request.user.is_valid():
-            return self.response_created(self.redirect(redirect_url))
-
-    def _check_new_user(self, init_password: bool = True) -> None:
-        if not self.request.user.first_time_login:
-            return
-        if self.request.session.get('confirmed') != 'yes':
-            return self.response_created(self.redirect('/agreement/'))
-        if init_password:
-            return self.response_created(self.redirect('/modpw/'))
-
-    def check_perm(self) -> None:
-        # TODO: 统一本函数和工具函数check_user_access
-        super().check_perm()
-        if self.check_access:
-            self._check_access()
-            self._check_new_user()
-
-    def dispatch_prepare(self, method: str):
-        return self.default_prepare(method, return_needed=False,
-                                    prepare_needed=self.need_prepare)
-
-    def get_logger(self):
-        return _ProfileLogger.getLogger(self.logger_name)
-
-
-class ProfileTemplateView(ProfileView, SecureTemplateView):
-    logger_name: str = 'ProfileError'
-    page_name: str
-
-    def render(self, **kwargs):
-        if not hasattr(self, 'page_name'):
-            raise _ImproperlyConfigured('page_name is not defined!')
-        self.extra_context['bar_display'] = utils.get_sidebar_and_navbar(
-            self.request.user, self.page_name)
-        return super().render(**kwargs)
-
-
-class ProfileJsonView(ProfileView, SecureJsonView):
-    logger_name: str = 'ProfileAPIerror'
-
-    def json_response(self, extra_data = None, **kwargs):
-        _ProfileLogger.getLogger('recording').info('json_response')
-        return super().json_response(extra_data, **kwargs)
+from app.view.dependency import *
